chat_management/requirements.txt
fastapi[all]
uvicorn[standard]
PyJWT==2.10.1
pydantic-settings